port over one card to conditionally render multiple sizes
